Refactor few elements
