+++ app/src/main/java/ru/practicum/android/diploma/di/DataModule.kt
@@ -5,7 +5,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.dsl.module
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import ru.practicum.android.diploma.search.data.SearchModelsConvertor
+import ru.practicum.android.diploma.search.data.convertor.SearchModelsConvertor
 import ru.practicum.android.diploma.search.data.network.NetworkClient
 import ru.practicum.android.diploma.search.data.network.RetrofitNetworkClient
 import ru.practicum.android.diploma.search.data.network.SearchAPI
