+++ app/src/main/java/ru/practicum/android/diploma/di/InteractorModule.kt
@@ -1,7 +1,7 @@
 package ru.practicum.android.diploma.di
 
 import org.koin.dsl.module
-import ru.practicum.android.diploma.search.domain.SearchInteractor
+import ru.practicum.android.diploma.search.domain.api.SearchInteractor
 import ru.practicum.android.diploma.search.domain.impl.SearchInteractorImpl
 import ru.practicum.android.diploma.sharing.domain.api.ResourceInteractor
 import ru.practicum.android.diploma.sharing.domain.api.SharingInteractor
