+++ app/src/main/java/ru/practicum/android/diploma/search/data/dto/SearchResponse.kt
@@ -1,7 +1,7 @@
 package ru.practicum.android.diploma.search.data.dto
 
 import com.google.gson.annotations.SerializedName
-import ru.practicum.android.diploma.search.data.dto.model.VacancyDTO
+import ru.practicum.android.diploma.search.data.model.VacancyDTO
 
 data class SearchResponse(
     @SerializedName("items") val vacancies: List<VacancyDTO>,
