--- app/src/main/java/ru/practicum/android/diploma/search/data/dto/model/AddressDTO.kt
@@ -1,3 +1,0 @@
-package ru.practicum.android.diploma.search.data.dto.model
-
-data class AddressDTO(val city: String?)
