+++ app/src/main/java/ru/practicum/android/diploma/search/data/impl/SearchRepositoryImpl.kt
@@ -2,15 +2,15 @@ package ru.practicum.android.diploma.search.data.impl
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import ru.practicum.android.diploma.search.data.Resource
-import ru.practicum.android.diploma.search.data.SearchModelsConvertor
+import ru.practicum.android.diploma.search.data.convertor.SearchModelsConvertor
 import ru.practicum.android.diploma.search.data.dto.SearchRequest
 import ru.practicum.android.diploma.search.data.dto.SearchResponse
 import ru.practicum.android.diploma.search.data.network.NetworkClient
-import ru.practicum.android.diploma.search.domain.SearchRepository
-import ru.practicum.android.diploma.search.domain.model.Vacancy
+import ru.practicum.android.diploma.search.domain.api.SearchRepository
+import ru.practicum.android.diploma.search.domain.model.ServerVacanciesResponse
 import ru.practicum.android.diploma.sharing.data.ResourceProvider
 import ru.practicum.android.diploma.util.Constants
+import ru.practicum.android.diploma.util.Resource
 
 class SearchRepositoryImpl(
     private val networkClient: NetworkClient,
@@ -20,15 +20,18 @@ class SearchRepositoryImpl(
     override suspend fun searchVacancies(
         expression: String,
         filters: Map<String, String>
-    ): Flow<Resource<List<Vacancy>>> = flow {
+    ): Flow<Resource<ServerVacanciesResponse>> = flow {
         val response = networkClient.doRequest(SearchRequest(expression, filters))
         when (response.result) {
             Constants.SUCCESS -> emit(
                 Resource.Success(
-                    (response as SearchResponse).vacancies.map { vacancyDTO -> convertor.map(vacancyDTO) },
-                    numberOfVacancies = response.numberOfVacancies
+                    ServerVacanciesResponse(
+                        (response as SearchResponse).vacancies.map { vacancyDTO -> convertor.map(vacancyDTO) },
+                        response.numberOfVacancies
+                    )
                 )
             )
+
             Constants.NOT_FOUND -> emit(Resource.Error(resourceProvider.getErrorEmptyListVacancy()))
             Constants.SERVER_ERROR -> emit(Resource.Error(resourceProvider.getErrorServer()))
             Constants.CONNECTION_ERROR -> emit(Resource.Error(resourceProvider.getErrorInternetConnection()))
