+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/AddressDTO.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.search.data.model
+
+data class AddressDTO(val city: String?)
