+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/EmployerDTO.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.search.data.dto.model
+package ru.practicum.android.diploma.search.data.model
 
 import com.google.gson.annotations.SerializedName
