--- app/src/main/java/ru/practicum/android/diploma/search/domain/SearchInteractor.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.search.domain
-
-import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.search.domain.model.Vacancy
-
-interface SearchInteractor {
-    suspend fun searchVacancies(expression: String, filters: Map<String, String>): Flow<Pair<List<Vacancy>?, String?>>
-    /* suspend fun getNumbersOfVacancies(expression: String, filters: Map<String, String>): Flow<Int?> */
-}
