+++ app/src/main/java/ru/practicum/android/diploma/search/domain/model/ServerVacanciesResponse.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.search.domain.model
+
+class ServerVacanciesResponse(
+    val vacancies: List<Vacancy?>?,
+    val numberOfVacancies: Int?,
+)
